+++ src/mnemonic.hpp
@@ -7,16 +7,20 @@
 #include "mnemonic.h"
 #include "mnemonic_ml.hpp"
 #include "mnemonic_bit.hpp"
-#include "mnemonic_branch.hpp"
 #include "mnemonic_calc.hpp"
 #include "mnemonic_data.hpp"
+#include "mnemonic_djnz.hpp"
 #include "mnemonic_ex.hpp"
 #include "mnemonic_im.hpp"
 #include "mnemonic_increment.hpp"
 #include "mnemonic_io.hpp"
+#include "mnemonic_jp.hpp"
+#include "mnemonic_jr.hpp"
 #include "mnemonic_load.hpp"
+#include "mnemonic_rst.hpp"
 #include "mnemonic_shift.hpp"
 #include "mnemonic_stack.hpp"
+#include "mnemonic_sub.hpp"
 #include "mnemonic_vgs.hpp"
 
 std::vector<TempAddr*> tempAddrs;
